style

+++ src/display2.c
@@ -811,7 +811,7 @@ void display2_af_spectre(uint_fast8_t x, uint_fast8_t y, dctx_t * pctx)
 			PACKEDCOLORMAIN_T * const fr = colmain_fb_draw();
 			static uint_fast8_t y_old_array [FIRBUFSIZE];
 
-			for (uint_fast16_t i = 3; i < afsp.w; i ++)
+			for (unsigned i = 3; i < afsp.w; i ++)
 			{
 				ASSERT(i < ARRAY_SIZE(y_old_array));
 				const uint_fast16_t fftpos = NORMALFFT - roundf(i / afsp.step);
@@ -819,7 +819,10 @@ void display2_af_spectre(uint_fast8_t x, uint_fast8_t y, dctx_t * pctx)
 				const FLOAT_t val = normalize(afsp.fft_buf [fftpos], 0, afsp.max_val, afsp.h);
 				const FLOAT_t yy = y_old_array [i] * (FLOAT_t) 0.8 + (FLOAT_t) 0.2 * val;
 				y_old_array [i] = yy;
-				colmain_line(fr, DIM_X, DIM_Y, afsp.x + i - 3, afsp.y - yy, afsp.x + i - 3, afsp.y, COLORMAIN_YELLOW, 0);
+				colmain_line(fr, DIM_X, DIM_Y,
+						afsp.x + i - 3, afsp.y - yy,
+						afsp.x + i - 3, afsp.y,
+						COLORMAIN_YELLOW, 0);
 			}
 		}
 		break;
